requirements.txt
youtube-transcript-api
pytube
moviepy
llama-index
python-dotenv
llama_hub